LPD-65557 Refactor. Move CreationMenu calculation code to the parent BaseSectionSystemFDSEntry class. Make it take httpServletRequest. This aligns with System Dataset APIs. No functional changes
