+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/frontend/data/set/BaseSectionSystemFDSEntry.java
@@ -5,23 +5,49 @@
 
 package com.liferay.site.cms.site.initializer.internal.frontend.data.set;
 
+import com.liferay.depot.constants.DepotConstants;
 import com.liferay.depot.model.DepotEntry;
+import com.liferay.depot.service.DepotEntryLocalService;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
+import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
 import com.liferay.info.constants.InfoDisplayWebKeys;
+import com.liferay.object.constants.ObjectDefinitionSettingConstants;
+import com.liferay.object.model.ObjectDefinitionSetting;
 import com.liferay.object.model.ObjectEntryFolder;
+import com.liferay.object.service.ObjectDefinitionSettingLocalService;
 import com.liferay.object.service.ObjectEntryFolderLocalServiceUtil;
+import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringUtil;
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.json.JSONUtil;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Group;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
+import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
 import jakarta.servlet.http.HttpServletRequest;
 
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * @author Daniel Sanz
+ * @author Marco Galluzzi
  */
 public abstract class BaseSectionSystemFDSEntry {
 
@@ -67,6 +93,61 @@ public abstract class BaseSectionSystemFDSEntry {
 		return sb.toString();
 	}
 
+	public CreationMenu getCreationMenu(HttpServletRequest httpServletRequest) {
+		return new CreationMenu() {
+			{
+				if (_hasAddEntryPermission(httpServletRequest)) {
+					for (DropdownItem dropdownItem :
+							getCreationMenuDropdownItems()) {
+
+						JSONArray depotEntriesJSONArray =
+							_getDepotEntriesJSONArray(
+								dropdownItem, httpServletRequest);
+
+						if (depotEntriesJSONArray == null) {
+							continue;
+						}
+
+						dropdownItem.putData(
+							"assetLibraries", depotEntriesJSONArray);
+
+						addPrimaryDropdownItem(dropdownItem);
+					}
+				}
+			}
+		};
+	}
+
+	public List<DropdownItem> getCreationMenuDropdownItems() {
+		return Collections.emptyList();
+	}
+
+	public void setDepotEntryLocalService(
+		DepotEntryLocalService depotEntryLocalService) {
+
+		this.depotEntryLocalService = depotEntryLocalService;
+	}
+
+	public void setGroupLocalService(GroupLocalService groupLocalService) {
+		this.groupLocalService = groupLocalService;
+	}
+
+	public void setObjectDefinitionSettingLocalService(
+		ObjectDefinitionSettingLocalService
+			objectDefinitionSettingLocalService) {
+
+		_objectDefinitionSettingLocalService =
+			objectDefinitionSettingLocalService;
+	}
+
+	public void setObjectEntryFolderModelResourcePermission(
+		ModelResourcePermission<ObjectEntryFolder>
+			objectEntryFolderModelResourcePermission) {
+
+		_objectEntryFolderModelResourcePermission =
+			objectEntryFolderModelResourcePermission;
+	}
+
 	protected String appendStatus(String filterString) {
 		return StringBundler.concat(
 			filterString, " and status in (", StringUtil.merge(_statuses, ", "),
@@ -74,6 +155,30 @@ public abstract class BaseSectionSystemFDSEntry {
 	}
 
 	protected abstract String getCMSSectionFilterString();
+
+	protected JSONArray getDepotEntriesJSONArray(
+		HttpServletRequest httpServletRequest) {
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		ObjectEntryFolder objectEntryFolder = getObjectEntryFolder(
+			themeDisplay.getCompanyId(),
+			httpServletRequest.getAttribute(InfoDisplayWebKeys.INFO_ITEM));
+
+		if (objectEntryFolder != null) {
+			return _getDepotEntriesJSONArray(
+				List.of(objectEntryFolder.getGroupId()), httpServletRequest);
+		}
+
+		return _getDepotEntriesJSONArray(
+			TransformUtil.transform(
+				depotEntryLocalService.getDepotEntries(
+					themeDisplay.getCompanyId(), DepotConstants.TYPE_SPACE),
+				DepotEntry::getGroupId),
+			httpServletRequest);
+	}
 
 	protected ObjectEntryFolder getObjectEntryFolder(
 		long companyId, Object object) {
@@ -97,9 +202,186 @@ public abstract class BaseSectionSystemFDSEntry {
 		return null;
 	}
 
+	protected DepotEntryLocalService depotEntryLocalService;
+	protected GroupLocalService groupLocalService;
+
+	private List<Long> _getAcceptedGroupIds(long objectDefinitionId) {
+		List<Long> acceptedGroupIds = new ArrayList<>();
+
+		ObjectDefinitionSetting objectDefinitionSetting =
+			_objectDefinitionSettingLocalService.fetchObjectDefinitionSetting(
+				objectDefinitionId,
+				ObjectDefinitionSettingConstants.NAME_ACCEPTED_GROUP_IDS);
+
+		for (String groupId :
+				StringUtil.split(objectDefinitionSetting.getValue())) {
+
+			DepotEntry depotEntry = depotEntryLocalService.fetchGroupDepotEntry(
+				GetterUtil.getLong(groupId));
+
+			if (depotEntry != null) {
+				acceptedGroupIds.add(depotEntry.getGroupId());
+			}
+		}
+
+		return acceptedGroupIds;
+	}
+
+	private JSONArray _getDepotEntriesJSONArray(
+		DropdownItem dropdownItem, HttpServletRequest httpServletRequest) {
+
+		Map<String, Object> dropdownItemData =
+			(HashMap<String, Object>)dropdownItem.get("data");
+
+		long objectDefinitionId = GetterUtil.getLong(
+			dropdownItemData.get("objectDefinitionId"));
+
+		if (objectDefinitionId != 0) {
+			return _getDepotEntriesJSONArray(
+				httpServletRequest, objectDefinitionId);
+		}
+
+		return getDepotEntriesJSONArray(httpServletRequest);
+	}
+
+	private JSONArray _getDepotEntriesJSONArray(
+		HttpServletRequest httpServletRequest, long objectDefinitionId) {
+
+		if (_isAcceptAllGroups(objectDefinitionId)) {
+			return getDepotEntriesJSONArray(httpServletRequest);
+		}
+
+		List<Long> acceptedGroupIds = _getAcceptedGroupIds(objectDefinitionId);
+
+		if (acceptedGroupIds.isEmpty()) {
+			return null;
+		}
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		ObjectEntryFolder objectEntryFolder = getObjectEntryFolder(
+			themeDisplay.getCompanyId(),
+			httpServletRequest.getAttribute(InfoDisplayWebKeys.INFO_ITEM));
+
+		if (objectEntryFolder != null) {
+			if (!acceptedGroupIds.contains(objectEntryFolder.getGroupId())) {
+				return null;
+			}
+
+			return _getDepotEntriesJSONArray(
+				List.of(objectEntryFolder.getGroupId()), httpServletRequest);
+		}
+
+		return _getDepotEntriesJSONArray(acceptedGroupIds, httpServletRequest);
+	}
+
+	private JSONArray _getDepotEntriesJSONArray(
+		List<Long> groupIds, HttpServletRequest httpServletRequest) {
+
+		JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
+
+		for (Long groupId : groupIds) {
+			JSONObject jsonObject = _getJSONObject(groupId, httpServletRequest);
+
+			if (jsonObject != null) {
+				jsonArray.put(jsonObject);
+			}
+		}
+
+		return jsonArray;
+	}
+
+	private JSONObject _getJSONObject(
+		long groupId, HttpServletRequest httpServletRequest) {
+
+		Group group = groupLocalService.fetchGroup(groupId);
+
+		if (group == null) {
+			return null;
+		}
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		return JSONUtil.put(
+			"externalReferenceCode", group.getExternalReferenceCode()
+		).put(
+			"groupId", group.getGroupId()
+		).put(
+			"name", group.getName(themeDisplay.getLocale())
+		);
+	}
+
+	private boolean _hasAddEntryPermission(
+		HttpServletRequest httpServletRequest) {
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		ObjectEntryFolder objectEntryFolder = getObjectEntryFolder(
+			themeDisplay.getCompanyId(),
+			httpServletRequest.getAttribute(InfoDisplayWebKeys.INFO_ITEM));
+
+		if (objectEntryFolder == null) {
+			return true;
+		}
+
+		try {
+			return _objectEntryFolderModelResourcePermission.contains(
+				themeDisplay.getPermissionChecker(),
+				objectEntryFolder.getObjectEntryFolderId(),
+				ActionKeys.ADD_ENTRY);
+		}
+		catch (PortalException portalException) {
+			if (_log.isDebugEnabled()) {
+				_log.debug(portalException);
+			}
+		}
+
+		return false;
+	}
+
+	private boolean _isAcceptAllGroups(long objectDefinitionId) {
+		ObjectDefinitionSetting objectDefinitionSetting =
+			_objectDefinitionSettingLocalService.fetchObjectDefinitionSetting(
+				objectDefinitionId,
+				ObjectDefinitionSettingConstants.NAME_ACCEPT_ALL_GROUPS);
+
+		if ((objectDefinitionSetting != null) &&
+			GetterUtil.getBoolean(objectDefinitionSetting.getValue())) {
+
+			return true;
+		}
+
+		objectDefinitionSetting =
+			_objectDefinitionSettingLocalService.fetchObjectDefinitionSetting(
+				objectDefinitionId,
+				ObjectDefinitionSettingConstants.NAME_ACCEPTED_GROUP_IDS);
+
+		if ((objectDefinitionSetting == null) ||
+			Validator.isNull(objectDefinitionSetting.getValue())) {
+
+			return true;
+		}
+
+		return false;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		BaseSectionSystemFDSEntry.class);
+
 	private static final List<Integer> _statuses = Arrays.asList(
 		WorkflowConstants.STATUS_APPROVED, WorkflowConstants.STATUS_DRAFT,
 		WorkflowConstants.STATUS_EXPIRED, WorkflowConstants.STATUS_PENDING,
 		WorkflowConstants.STATUS_SCHEDULED);
+
+	private ObjectDefinitionSettingLocalService
+		_objectDefinitionSettingLocalService;
+	private ModelResourcePermission<ObjectEntryFolder>
+		_objectEntryFolderModelResourcePermission;
 
 }
