+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/view_space_files_summary.jsp
@@ -23,7 +23,7 @@ ViewSpaceFilesSummarySectionDisplayContext viewSpaceFilesSummarySectionDisplayCo
 		<frontend-data-set:headless-display
 			additionalProps="<%= viewSpaceFilesSummarySectionDisplayContext.getAdditionalProps() %>"
 			apiURL="<%= viewSpaceFilesSummarySectionDisplayContext.getAPIURL() %>"
-			creationMenu="<%= viewSpaceFilesSummarySectionDisplayContext.getCreationMenu() %>"
+			creationMenu="<%= viewSpaceFilesSummarySectionDisplayContext.getCreationMenu(request) %>"
 			emptyState="<%= viewSpaceFilesSummarySectionDisplayContext.getEmptyState() %>"
 			fdsActionDropdownItems="<%= viewSpaceFilesSummarySectionDisplayContext.getFDSActionDropdownItems() %>"
 			formName="fm"
